refactor(app): split dashboard request planning, execution, and materialization
